Removed Account and Role class
Removed Account and Role class and then added their attributes to the User class
