+++ src/medfx/User.java
@@ -2,18 +2,29 @@ package medfx;
 
 public class User {
 	private PersonalInformation personalInfo;	//	Class holds the first name, last name, birthday, address, phone, email
-	private Account account;	//	username, password, role, changeUsername, changePassword
+	private String role;
+	private String username;
 
-	public User(PersonalInformation personalInfo, Account accout) {
+	public User(PersonalInformation personalInfo, String role) {
 		this.personalInfo = personalInfo;
-		this.account = account;
+		this.role = role;
+		
+		// generates username from personal info
+		this.username = personalInfo.getLastName() + "-" + personalInfo.getFirstName() + "-" + personalInfo.getDate();
+		
 	}
 	
 	public PersonalInformation getPersonalInfo() {
 		return this.personalInfo;
 	}
+
 	
-	public Account getAccount() {
-		return this.account;
+	public String getUsername() {
+		return username;
+	}
+	
+	
+	public String getRole() {
+		return role;
 	}
 }
